// src/modules/inventory-control/src/domain/enums/estoque-tipo-movimento.enum.ts
export enum EstoqueTipoMovimento {
  ENTRADA_COMPRA = 'ENTRADA_COMPRA',
  SAIDA_VENDA = 'SAIDA_VENDA',
  ENTRADA_DEVOLUCAO = 'ENTRADA_DEVOLUCAO',
  SAIDA_DEVOLUCAO = 'SAIDA_DEVOLUCAO',
  AJUSTE_POSITIVO = 'AJUSTE_POSITIVO',
  AJUSTE_NEGATIVO = 'AJUSTE_NEGATIVO',
  TRANSFERENCIA_ENTRADA = 'TRANSFERENCIA_ENTRADA',
  TRANSFERENCIA_SAIDA = 'TRANSFERENCIA_SAIDA',
  PRODUCAO_ENTRADA = 'PRODUCAO_ENTRADA',
  PRODUCAO_CONSUMO = 'PRODUCAO_CONSUMO',
  CONSUMO_INTERNO = 'CONSUMO_INTERNO',
  PERDA = 'PERDA',
  AVARIA = 'AVARIA',
  INVENTARIO = 'INVENTARIO',
}